Wifiology API Auth
This is a relatively big commit, but I have added both user auth
and API key auth support to the API (the only thing left to do
potentially is to add OAuth2 support). All of the new endpoints
should work and both basic and api key auth are tested now.